Do not launch preprocessing under transaction
